Extract into DiscoveryManager
