+++ common/src/main/java/com/wynntils/core/webapi/WebManager.java
@@ -4,22 +4,16 @@
  */
 package com.wynntils.core.webapi;
 
-import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.managers.CoreManager;
 import com.wynntils.core.webapi.request.RequestBuilder;
 import com.wynntils.core.webapi.request.RequestHandler;
 import com.wynntils.mc.event.WebSetupEvent;
-import com.wynntils.wynn.model.discoveries.objects.DiscoveryInfo;
-import com.wynntils.wynn.objects.profiles.DiscoveryProfile;
 import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.Type;
 import java.net.URL;
 import java.net.URLConnection;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 /** Provides and loads web content on demand */
@@ -32,10 +26,6 @@ public final class WebManager extends CoreManager {
     private static final RequestHandler handler = new RequestHandler();
 
     public static WebReader apiUrls = null;
-
-    private static final Gson gson = new Gson();
-
-    private static List<DiscoveryInfo> discoveryInfoList = new ArrayList<>();
 
     private static final String USER_AGENT = String.format(
             "Wynntils Artemis\\%s (%s) %s",
@@ -70,23 +60,6 @@ public final class WebManager extends CoreManager {
         handler.dispatch(async);
     }
 
-    public static void updateDiscoveries() {
-        if (apiUrls == null) return;
-
-        String url = apiUrls.get("Discoveries");
-        handler.addRequest(new RequestBuilder(url, "discoveries")
-                .cacheTo(new File(API_CACHE_ROOT, "discoveries.json"))
-                .handleJsonArray(discoveriesJson -> {
-                    Type type = new TypeToken<ArrayList<DiscoveryProfile>>() {}.getType();
-
-                    List<DiscoveryProfile> discoveries = gson.fromJson(discoveriesJson, type);
-                    discoveryInfoList =
-                            discoveries.stream().map(DiscoveryInfo::new).toList();
-                    return true;
-                })
-                .build());
-    }
-
     public static URLConnection generateURLRequest(String url) throws IOException {
         URLConnection st = new URL(url).openConnection();
         st.setRequestProperty("User-Agent", USER_AGENT);
@@ -101,10 +74,6 @@ public final class WebManager extends CoreManager {
         if (apiUrls == null) return null;
 
         return apiUrls.get(key);
-    }
-
-    public static List<DiscoveryInfo> getDiscoveryInfoList() {
-        return discoveryInfoList;
     }
 
     public static String getUserAgent() {
