+++ common/src/main/java/com/wynntils/wynn/model/discoveries/DiscoveryManager.java
@@ -4,6 +4,8 @@
  */
 package com.wynntils.wynn.model.discoveries;
 
+import com.google.common.reflect.TypeToken;
+import com.google.gson.Gson;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.managers.CoreManager;
 import com.wynntils.core.webapi.WebManager;
@@ -21,7 +23,11 @@ import com.wynntils.wynn.event.WorldStateEvent;
 import com.wynntils.wynn.model.CompassModel;
 import com.wynntils.wynn.model.discoveries.objects.DiscoveryInfo;
 import com.wynntils.wynn.model.discoveries.objects.DiscoveryType;
+import com.wynntils.wynn.objects.profiles.DiscoveryProfile;
 import com.wynntils.wynn.objects.profiles.TerritoryProfile;
+import java.io.File;
+import java.lang.reflect.Type;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 import net.minecraft.ChatFormatting;
@@ -32,9 +38,11 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 public class DiscoveryManager extends CoreManager {
     private static final DiscoveryContainerQueries CONTAINER_QUERIES = new DiscoveryContainerQueries();
+    private static final Gson GSON = new Gson();
 
     private static List<DiscoveryInfo> discoveries = List.of();
     private static List<DiscoveryInfo> secretDiscoveries = List.of();
+    private static List<DiscoveryInfo> discoveryInfoList = new ArrayList<>();
 
     private static List<Component> discoveriesTooltip = List.of();
     private static List<Component> secretDiscoveriesTooltip = List.of();
@@ -84,7 +92,7 @@ public class DiscoveryManager extends CoreManager {
         Utils.openUrl(wikiUrl);
     }
 
-    public static void queryDiscoveries() {
+    private static void queryDiscoveries() {
         CONTAINER_QUERIES.queryDiscoveries();
     }
 
@@ -116,6 +124,10 @@ public class DiscoveryManager extends CoreManager {
 
     public static Stream<DiscoveryInfo> getAllDiscoveries() {
         return Stream.concat(discoveries.stream(), secretDiscoveries.stream());
+    }
+
+    public static List<DiscoveryInfo> getDiscoveryInfoList() {
+        return discoveryInfoList;
     }
 
     private static void locateSecretDiscovery(String name, DiscoveryOpenAction action) {
@@ -187,6 +199,28 @@ public class DiscoveryManager extends CoreManager {
         RequestHandler handler = new RequestHandler();
 
         handler.addAndDispatch(query, true);
+    }
+
+    private static void updateDiscoveriesResource() {
+        if (WebManager.apiUrls == null) return;
+
+        String url = WebManager.apiUrls.get("Discoveries");
+        WebManager.getHandler().addRequest(new RequestBuilder(url, "discoveries")
+                .cacheTo(new File(WebManager.API_CACHE_ROOT, "discoveries.json"))
+                .handleJsonArray(discoveriesJson -> {
+                    Type type = new TypeToken<ArrayList<DiscoveryProfile>>() {}.getType();
+
+                    List<DiscoveryProfile> discoveries = GSON.fromJson(discoveriesJson, type);
+                    discoveryInfoList =
+                            discoveries.stream().map(DiscoveryInfo::new).toList();
+                    return true;
+                })
+                .build());
+    }
+
+    public static void reloadDiscoveries() {
+        updateDiscoveriesResource();
+        queryDiscoveries();
     }
 
     public enum DiscoveryOpenAction {
